Revert "UNDO fixes to reasonVFTableBelongsToClass that were already made public."
This reverts commit 07812a87717b8bd978efb47580e6bdb723b52a05.

This commit was based on several attempts to fix it, as documented in rules.cpp of the model compiler:

Attempt 1: Check for embedded classes instead of a lack of inherited classes.
Attempt 2: Check if the offset in question falls inside an embedded object
Attempt 3: Check for a non-virt ancestor class at offset, or no embedded class at offset 0.
Attempt 4: Check for a non-virt ancestor class at offset.  If there is one that has a vftable belonging at (its) offset 0, then we own the vftable.  Or if the offset is 0 and there is no embedded class at offset 0, we own the vftable.
Attempt 5: Checking for a virt ancestor at offset is ok too.

Note: Attempt 5 does not currently have an equivalent implementation in Prolog because factDerivedClass does not have a mechanism for distinguishing virtual bases.