Soundness: Fix empty base bug in reasonNOTConstructor_G
The rule observes that if there is a vftable install at offset 0 and there is a call to a method at offset 0, then the callee must install a vftable at offset 0 or it is not a constructor.  Unfortunately, if there is an empty base at offset 0 this is not true, as the empty base would have no vfptr slots and yet could still have a constructor run on offset 0.